requirements.txt
python-dotenv==1.0.1
loguru==0.7.2
requests==2.31.0
python-multipart==0.0.9
aiohttp==3.9.3
tenacity==8.2.3 

fastapi~=0.110.0
uvicorn~=0.30.0
python-dotenv==1.0.1
pydantic~=2.11.5